Updated build files

+++ build_macos.sh
@@ -2,6 +2,19 @@
 # Stop script on errors.
 set -e
 
+NO_INTERACTIVE=false
+
+# Parse command line arguments
+for arg in "$@"
+do
+    case $arg in
+        --no-interactive)
+        NO_INTERACTIVE=true
+        shift
+        ;;
+    esac
+done
+
 # Check build time dependencies.
 if ! command -v python3 &> /dev/null
 then
@@ -29,7 +42,13 @@ export WORK_DIR=$PROJECT_DIR/Mediasoup/dependencies
 echo "WORK_DIR = $WORK_DIR"
 export BUILD_DIR=$(pwd)/build
 echo "BUILD_DIR = $BUILD_DIR"
-export OUTPUT_DIR=$(pwd)/bin
+
+if [ "$NO_INTERACTIVE" = false ]; then
+    export OUTPUT_DIR=$(pwd)/bin
+else
+    export OUTPUT_DIR=$(pwd)/bin/mac
+fi
+
 echo "OUTPUT_DIR = $OUTPUT_DIR"
 export PATCHES_DIR=$(pwd)/patches
 echo "PATCHES_DIR = $PATCHES_DIR"
@@ -57,26 +76,29 @@ function clearArtifacts() {
     echo 'BUILD_DIR created'
 }
 
-while true
-do
-    read -n 1 -p "Clear old build artifacts? (Y|n): " INPUT_STRING
-    # echo ""
-    case $INPUT_STRING in
-        n|N)
-            echo ""
-            break
-            ;;
-        y|Y|"")
-            echo ""
-            clearArtifacts
-            break
-            ;;
-        *)
-            echo -ne "\r\033[0K\r"
-            tput bel
-            ;;
-    esac
-done
+if [ "$NO_INTERACTIVE" = true ]; then
+    clearArtifacts
+else
+    while true
+    do
+        read -n 1 -p "Clear old build artifacts? (Y|n): " INPUT_STRING
+        case $INPUT_STRING in
+            n|N)
+                echo ""
+                break
+                ;;
+            y|Y|"")
+                echo ""
+                clearArtifacts
+                break
+                ;;
+            *)
+                echo -ne "\r\033[0K\r"
+                tput bel
+                ;;
+        esac
+    done
+fi
 
 function refetchLibmediasoupclient() {
     echo 'Cloning libmediasoupclient'
@@ -88,25 +110,27 @@ function refetchLibmediasoupclient() {
 if [ -d $WORK_DIR/libmediasoupclient ]
 then
     echo "libmediasoupclient is already on disk"
-    while true
-    do
-        read -n 1 -p "Refetch libmediasoupclient (y|N): " INPUT_STRING
-        case $INPUT_STRING in
-            n|N|"")
-                echo ""
-                break
-                ;;
-            y|Y)
-                echo ""
-                refetchLibmediasoupclient
-                break
-                ;;
-            *)
-                echo -ne "\r\033[0K\r"
-                tput bel
-                ;;
-        esac
-    done
+    if [ "$NO_INTERACTIVE" = false ]; then
+        while true
+        do
+            read -n 1 -p "Refetch libmediasoupclient (y|N): " INPUT_STRING
+            case $INPUT_STRING in
+                n|N|"")
+                    echo ""
+                    break
+                    ;;
+                y|Y)
+                    echo ""
+                    refetchLibmediasoupclient
+                    break
+                    ;;
+                *)
+                    echo -ne "\r\033[0K\r"
+                    tput bel
+                    ;;
+            esac
+        done
+    fi
 else
     refetchLibmediasoupclient
 fi
@@ -121,23 +145,25 @@ function refetchDepotTools() {
 if [ -d $WORK_DIR/depot_tools ]
 then
     echo "depot_tools is already on disk"
-    while true
-    do
-        read -n 1 -p "Refetch depot_tools (y|N): " INPUT_STRING
-        echo ""
-        case $INPUT_STRING in
-            n|N|"")
-                break
-                ;;
-            y|Y)
-                refetchDepotTools
-                break
-                ;;
-            *)
-                tput bel
-                ;;
-        esac
-    done
+    if [ "$NO_INTERACTIVE" = false ]; then
+        while true
+        do
+            read -n 1 -p "Refetch depot_tools (y|N): " INPUT_STRING
+            echo ""
+            case $INPUT_STRING in
+                n|N|"")
+                    break
+                    ;;
+                y|Y)
+                    refetchDepotTools
+                    break
+                    ;;
+                *)
+                    tput bel
+                    ;;
+            esac
+        done
+    fi
 else
     refetchDepotTools
 fi
@@ -161,6 +187,7 @@ function patchWebRTC() {
     patch -b -p0 -d $WORK_DIR < $PATCHES_DIR/objc_audio_device_module_h.patch
     patch -b -p0 -d $WORK_DIR < $PATCHES_DIR/objc_audio_device_module_mm.patch
     patch -b -p0 -d $WORK_DIR < $PATCHES_DIR/absl_threadlocal.patch
+    patch -b -p0 -d $WORK_DIR < $PATCHES_DIR/task_factory.patch
 }
 
 function refetchWebRTC() {
@@ -216,29 +243,34 @@ function resetWebRTC() {
 if [ -d $WORK_DIR/webrtc ]
 then
     echo "WebRTC is already on disk"
-    while true
-    do
-        read -n 1 -p "Refetch WebRTC? (f)ull clone | (r)eset local changes | (N)o: " INPUT_STRING
-        echo ""
-        case $INPUT_STRING in
-            n|N|"")
-                break
-                ;;
-            f|F)
-                refetchWebRTC
-                patchWebRTC
-                break
-                ;;
-            r|R)
-                resetWebRTC
-                patchWebRTC
-                break
-                ;;
-            *)
-                tput bel
-                ;;
-        esac
-    done
+    if [ "$NO_INTERACTIVE" = false ]; then
+        while true
+        do
+            read -n 1 -p "Refetch WebRTC? (f)ull clone | (r)eset local changes | (N)o: " INPUT_STRING
+            echo ""
+            case $INPUT_STRING in
+                n|N|"")
+                    break
+                    ;;
+                f|F)
+                    refetchWebRTC
+                    patchWebRTC
+                    break
+                    ;;
+                r|R)
+                    resetWebRTC
+                    patchWebRTC
+                    break
+                    ;;
+                *)
+                    tput bel
+                    ;;
+            esac
+        done
+    else
+        resetWebRTC
+        patchWebRTC
+    fi
 else
     refetchWebRTC
     patchWebRTC
@@ -332,11 +364,15 @@ echo 'Finish lipo'
 
 echo 'Create xcframework'
 cd $BUILD_DIR/WebRTC
-rm -rf $OUTPUT_DIR/WebRTC.xcframework
 
-xcodebuild -create-xcframework \
-    -framework mac/WebRTC.framework \
-    -output $OUTPUT_DIR/WebRTC.xcframework
+if [ "$NO_INTERACTIVE" = false ]; then
+    rm -rf $OUTPUT_DIR/WebRTC.xcframework
+    xcodebuild -create-xcframework \
+        -framework mac/WebRTC.framework \
+        -output $OUTPUT_DIR/WebRTC.xcframework
+else
+    mv mac/WebRTC.framework $OUTPUT_DIR/WebRTC.framework
+fi
 
 echo "finish building webrtc"
 
@@ -394,41 +430,53 @@ function rebuildLMSC() {
         $BUILD_DIR/libmediasoupclient/mac/x64/libmediasoupclient/libsdptransform/libsdptransform.a \
         -output $BUILD_DIR/libmediasoupclient/mac/fat/libsdptransform.a
 
-    echo "create mediasoupclient.xcframework"
-    xcodebuild -create-xcframework \
-        -library $BUILD_DIR/libmediasoupclient/mac/fat/libmediasoupclient.a \
-        -output $OUTPUT_DIR/mediasoupclient.xcframework
-      echo " create sdptransform.xcframework"
-    xcodebuild -create-xcframework \
-        -library $BUILD_DIR/libmediasoupclient/mac/fat/libsdptransform.a \
-        -output $OUTPUT_DIR/sdptransform.xcframework
+
+    if [ "$NO_INTERACTIVE" = false ]; then
+        echo "create mediasoupclient.xcframework"
+        xcodebuild -create-xcframework \
+            -library $BUILD_DIR/libmediasoupclient/mac/fat/libmediasoupclient.a \
+            -output $OUTPUT_DIR/mediasoupclient.xcframework
+        echo " create sdptransform.xcframework"
+        xcodebuild -create-xcframework \
+            -library $BUILD_DIR/libmediasoupclient/mac/fat/libsdptransform.a \
+            -output $OUTPUT_DIR/sdptransform.xcframework
+    else
+        mv $BUILD_DIR/libmediasoupclient/mac/fat/libmediasoupclient.a $OUTPUT_DIR/libmediasoupclient.a
+        mv $BUILD_DIR/libmediasoupclient/mac/fat/libsdptransform.a $OUTPUT_DIR/libsdptransform.a
+    fi
     echo "finish"
 }
 
 if [ -d $BUILD_DIR/libmediasoupclient ]
 then
     echo "libmediasoupclient is already built"
-    while true
-    do
-        read -n 1 -p "Rebuild libmediasoupclient (y|N): " INPUT_STRING
-        case $INPUT_STRING in
-            n|N|"")
-                echo ""
-                break
-                ;;
-            y|Y)
-                echo ""
-                rebuildLMSC
-                break
-                ;;
-            *)
-                echo -ne "\r\033[0K\r"н
-                tput bel
-                ;;
-        esac
-    done
+    if [ "$NO_INTERACTIVE" = false ]; then
+        while true
+        do
+            read -n 1 -p "Rebuild libmediasoupclient (y|N): " INPUT_STRING
+            case $INPUT_STRING in
+                n|N|"")
+                    echo ""
+                    break
+                    ;;
+                y|Y)
+                    echo ""
+                    rebuildLMSC
+                    break
+                    ;;
+                *)
+                    echo -ne "\r\033[0K\r"н
+                    tput bel
+                    ;;
+            esac
+        done
+    else
+        rebuildLMSC
+    fi
 else
     rebuildLMSC
 fi
 
-open $PROJECT_DIR/Mediasoup.xcodeproj
+if [ "$NO_INTERACTIVE" = false ]; then
+    open $PROJECT_DIR/Mediasoup.xcodeproj
+fi
